Create findthenextperfectsquare7kyu
The method findNextSquare takes a long integer sq as input.
It calculates the square root of sq using Math.sqrt() and casts it to a long. It increments the square root by 1 and stores it in next.
If the square of the calculated square root is equal to the original number sq, it means sq is a perfect square. In this case, it returns the square of next. If sq is not a perfect square, it returns -1.